Google ADK/Agente/requirements.txt
psycopg2==2.9.11
psycopg-binary==3.3.3
langchain==1.2.10
langchain-postgres==0.0.17
langchain-google-genai==4.2.1
dotenv==0.9.9
google-adk==1.28.1
google-genai==1.65.0
streamlit==1.54.0
azure-functions==1.24.0
azure-functions-durable==1.5.0
sqlalchemy==2.0.46